% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hBayesDMregression-package.R
\docType{package}
\name{hBayesDMregression-package}
\alias{hBayesDMregression}
\alias{hBayesDMregression-package}
\title{hBayesDMregression: Regression of Iowa Gambling Task parameters on covariates}
\description{
Perform hierarchical linear regression of Iowa Gambling Task model parameters on a set of user-supplied covariates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamoshen/hBayesDMregression}
  \item Report bugs at \url{https://github.com/adamoshen/hBayesDMregression/issues}
}

}
\author{
\strong{Maintainer}: Adam Shen \email{adam.shen@carleton.ca}

Authors:
\itemize{
  \item Dave Campbell \email{davecampbell@math.carleton.ca} [funder, reviewer]
}

}
\keyword{internal}
